✨ Feat: mapCategory hook 통일
